requirements.txt
langchain-experimental
langchain-openai
python-dotenv
streamlit
langchain
tabulate
pydantic
pandas